HACK: Abort if we do not receive a message from malamute for 10 minutes
This is a band-aid for 1.4, where we have issues that the connection
with malamute breaks for some reason. Normally, we should be receiving
metrics from fty-nut and others each 30 seconds. If we receive nothing
for 10 minutes, it is a clear indication that something went wrong. Stop
the agent and let systemd start it anew.